moved more files inside the client folder
