+++ cmd/main.go
@@ -8,14 +8,12 @@ import (
 	"github.com/LiU-SeeGoals/controller/internal/client"
 	"github.com/LiU-SeeGoals/controller/internal/config"
 	"github.com/LiU-SeeGoals/controller/internal/gamestate"
-	"github.com/LiU-SeeGoals/controller/internal/receiver"
 	"github.com/LiU-SeeGoals/controller/internal/simulator"
-	"github.com/LiU-SeeGoals/controller/internal/webserver"
 )
 
 func main() {
 	gs := gamestate.NewGameState(3)
-	ssl_receiver := receiver.NewSSLReceiver(config.GetSSLClientAddress())
+	ssl_receiver := client.NewSSLVisionClient(config.GetSSLClientAddress())
 
 	ai_blue := ai.NewAi(gamestate.Blue)
 	sim_client_blue := client.NewSimClient(config.GetSimBlueTeamAddress())
@@ -43,6 +41,6 @@ func main() {
 
 		terminal_messages := []string{fmt.Sprintf("Blue score: %.2f AnticipatedScore: %.2f", score_blue, antBlue), fmt.Sprintf("Yellow score: %.2f AnticipatedScore: %.2f", score_yellow, antYellow)}
 
-		webserver.UpdateWebGUI(gs, blue_actions, terminal_messages)
+		client.UpdateWebGUI(gs, blue_actions, terminal_messages)
 	}
 }
