+++ examples/example_ssl_vision.go
@@ -1,54 +1,55 @@
 package examples
 
-import (
-	"fmt"
+// import (
+// 	"fmt"
 
-	"github.com/LiU-SeeGoals/controller/internal/receiver"
-	"github.com/LiU-SeeGoals/proto_go/ssl_vision"
-)
+// 	"github.com/LiU-SeeGoals/controller/internal/client"
 
-func SSLVisionExample() {
-	var packet ssl_vision.SSL_WrapperPacket
-	packetChan := setupSSLVisionReceiver()
+// 	"github.com/LiU-SeeGoals/proto_go/ssl_vision"
+// )
 
-	var detect *ssl_vision.SSL_DetectionFrame = nil
-	for {
-		packet = <-packetChan
+// func SSLVisionExample() {
+// 	var packet ssl_vision.SSL_WrapperPacket
+// 	packetChan := setupSSLVisionReceiver()
 
-		detect = packet.GetDetection()
+// 	var detect *ssl_vision.SSL_DetectionFrame = nil
+// 	for {
+// 		packet = <-packetChan
 
-		fmt.Println("-- Blue team --")
-		for _, robot := range detect.GetRobotsBlue() {
-			fmt.Printf("id: %v \n", robot.GetRobotId())
-			fmt.Printf("x: %v \n", robot.GetX())
-			fmt.Printf("y: %v \n", robot.GetY())
-		}
+// 		detect = packet.GetDetection()
 
-		fmt.Println("-- Yellow team --")
-		for _, robot := range detect.GetRobotsYellow() {
-			fmt.Printf("id: %v \n", robot.GetRobotId())
-			fmt.Printf("x: %v \n", robot.GetX())
-			fmt.Printf("y: %v \n", robot.GetY())
-		}
+// 		fmt.Println("-- Blue team --")
+// 		for _, robot := range detect.GetRobotsBlue() {
+// 			fmt.Printf("id: %v \n", robot.GetRobotId())
+// 			fmt.Printf("x: %v \n", robot.GetX())
+// 			fmt.Printf("y: %v \n", robot.GetY())
+// 		}
 
-		fmt.Println("-- Ball --")
-		for _, ball := range detect.GetBalls() {
-			fmt.Printf("x: %v \n", ball.GetX())
-			fmt.Printf("y: %v \n", ball.GetY())
-			fmt.Printf("z: %v \n", ball.GetZ())
-		}
-	}
+// 		fmt.Println("-- Yellow team --")
+// 		for _, robot := range detect.GetRobotsYellow() {
+// 			fmt.Printf("id: %v \n", robot.GetRobotId())
+// 			fmt.Printf("x: %v \n", robot.GetX())
+// 			fmt.Printf("y: %v \n", robot.GetY())
+// 		}
 
-}
+// 		fmt.Println("-- Ball --")
+// 		for _, ball := range detect.GetBalls() {
+// 			fmt.Printf("x: %v \n", ball.GetX())
+// 			fmt.Printf("y: %v \n", ball.GetY())
+// 			fmt.Printf("z: %v \n", ball.GetZ())
+// 		}
+// 	}
 
-// Start a SSL Vision receiver, returns a channel from
-// which SSL wrapper packets can be obtained.
-func setupSSLVisionReceiver() chan ssl_vision.SSL_WrapperPacket {
-	recv := receiver.NewSSLReceiver("224.5.23.2:10020")
-	recv.Connect()
+// }
 
-	c := make(chan ssl_vision.SSL_WrapperPacket)
-	go recv.Receive(c)
+// // Start a SSL Vision receiver, returns a channel from
+// // which SSL wrapper packets can be obtained.
+// func setupSSLVisionReceiver() chan ssl_vision.SSL_WrapperPacket {
+// 	recv := client.NewSSLVisionClient("224.5.23.2:10020")
+// 	recv.Connect()
 
-	return c
-}
+// 	c := make(chan ssl_vision.SSL_WrapperPacket)
+// 	go recv.Receive(c)
+
+// 	return c
+// }
