--- internal/client/test_client.go
@@ -1,1 +1,0 @@
-package client
